+++ src/commands/impls/server.rs
@@ -35,6 +35,8 @@ pub async fn quit<C: ClientLike>(client: &C) -> Result<(), RedisError> {
   let _ = client.send_command(command)?;
   let _ = rx.await??;
   utils::abort_network_timeout_task(&inner);
+  inner.notifications.close_public_receivers();
+  inner.backchannel.write().await.check_and_disconnect(&inner, None).await;
 
   Ok(())
 }
@@ -62,6 +64,8 @@ pub async fn shutdown<C: ClientLike>(client: &C, flags: Option<ShutdownFlags>) -
   let _ = client.send_command(command)?;
   let _ = rx.await??;
   utils::abort_network_timeout_task(&inner);
+  inner.notifications.close_public_receivers();
+  inner.backchannel.write().await.check_and_disconnect(&inner, None).await;
 
   Ok(())
 }