+++ src/protocol/command.rs
@@ -155,12 +155,15 @@ pub enum RedisCommandKind {
   ClientKill,
   ClientList,
   ClientGetName,
-  ClientGetRedir,
   ClientPause,
   ClientUnpause,
   ClientUnblock,
   ClientReply,
   ClientSetname,
+  ClientGetRedir,
+  ClientTracking,
+  ClientTrackingInfo,
+  ClientCaching,
   ClusterAddSlots,
   ClusterCountFailureReports,
   ClusterCountKeysInSlot,
@@ -402,6 +405,11 @@ pub enum RedisCommandKind {
   FunctionLoad,
   FunctionRestore,
   FunctionStats,
+  PubsubChannels,
+  PubsubNumpat,
+  PubsubNumsub,
+  PubsubShardchannels,
+  PubsubShardnumsub,
   // Commands with custom state or commands that don't map directly to the server's command interface.
   _Hello(RespVersion),
   _AuthAllCluster,
@@ -414,6 +422,8 @@ pub enum RedisCommandKind {
   _FunctionFlushCluster,
   _FunctionDeleteCluster,
   _FunctionRestoreCluster,
+  // When in RESP3 mode and **not** using the `bcast` arg then we send the command on all cluster node connections
+  _ClientTrackingCluster,
   _Custom(CustomCommand),
 }
 
@@ -570,13 +580,16 @@ impl RedisCommandKind {
       RedisCommandKind::ClientInfo => "CLIENT INFO",
       RedisCommandKind::ClientKill => "CLIENT KILL",
       RedisCommandKind::ClientList => "CLIENT LIST",
-      RedisCommandKind::ClientGetRedir => "CLIENT GETREDIR",
       RedisCommandKind::ClientGetName => "CLIENT GETNAME",
       RedisCommandKind::ClientPause => "CLIENT PAUSE",
       RedisCommandKind::ClientUnpause => "CLIENT UNPAUSE",
       RedisCommandKind::ClientUnblock => "CLIENT UNBLOCK",
       RedisCommandKind::ClientReply => "CLIENT REPLY",
       RedisCommandKind::ClientSetname => "CLIENT SETNAME",
+      RedisCommandKind::ClientGetRedir => "CLIENT GETREDIR",
+      RedisCommandKind::ClientTracking => "CLIENT TRACKING",
+      RedisCommandKind::ClientTrackingInfo => "CLIENT TRACKINGINFO",
+      RedisCommandKind::ClientCaching => "CLIENT CACHING",
       RedisCommandKind::ClusterAddSlots => "CLUSTER ADDSLOTS",
       RedisCommandKind::ClusterCountFailureReports => "CLUSTER COUNT-FAILURE-REPORTS",
       RedisCommandKind::ClusterCountKeysInSlot => "CLUSTER COUNTKEYSINSLOT",
@@ -819,6 +832,7 @@ impl RedisCommandKind {
       RedisCommandKind::_FunctionFlushCluster => "FUNCTION FLUSH CLUSTER",
       RedisCommandKind::_FunctionDeleteCluster => "FUNCTION DELETE CLUSTER",
       RedisCommandKind::_FunctionRestoreCluster => "FUNCTION RESTORE CLUSTER",
+      RedisCommandKind::_ClientTrackingCluster => "CLIENT TRACKING CLUSTER",
       RedisCommandKind::Fcall => "FCALL",
       RedisCommandKind::FcallRO => "FCALL_RO",
       RedisCommandKind::FunctionDelete => "FUNCTION DELETE",
@@ -829,6 +843,11 @@ impl RedisCommandKind {
       RedisCommandKind::FunctionLoad => "FUNCTION LOAD",
       RedisCommandKind::FunctionRestore => "FUNCTION RESTORE",
       RedisCommandKind::FunctionStats => "FUNCTION STATS",
+      RedisCommandKind::PubsubChannels => "PUBSUB CHANNELS",
+      RedisCommandKind::PubsubNumpat => "PUBSUB NUMPAT",
+      RedisCommandKind::PubsubNumsub => "PUBSUB NUMSUB",
+      RedisCommandKind::PubsubShardchannels => "PUBSUB SHARDCHANNELS",
+      RedisCommandKind::PubsubShardnumsub => "PUBSUB SHARDNUMSUB",
       RedisCommandKind::_Custom(ref kind) => &kind.cmd,
     }
   }
@@ -871,12 +890,15 @@ impl RedisCommandKind {
       | RedisCommandKind::ClientKill
       | RedisCommandKind::ClientList
       | RedisCommandKind::ClientGetName
-      | RedisCommandKind::ClientGetRedir
       | RedisCommandKind::ClientPause
       | RedisCommandKind::ClientUnpause
       | RedisCommandKind::ClientUnblock
       | RedisCommandKind::ClientReply
-      | RedisCommandKind::ClientSetname => "CLIENT",
+      | RedisCommandKind::ClientSetname
+      | RedisCommandKind::ClientCaching
+      | RedisCommandKind::ClientTrackingInfo
+      | RedisCommandKind::ClientTracking
+      | RedisCommandKind::ClientGetRedir => "CLIENT",
       RedisCommandKind::ClusterAddSlots
       | RedisCommandKind::ClusterCountFailureReports
       | RedisCommandKind::ClusterCountKeysInSlot
@@ -1125,8 +1147,14 @@ impl RedisCommandKind {
       | RedisCommandKind::_FunctionRestoreCluster
       | RedisCommandKind::_FunctionDeleteCluster
       | RedisCommandKind::_FunctionLoadCluster => "FUNCTION",
+      RedisCommandKind::PubsubChannels
+      | RedisCommandKind::PubsubNumpat
+      | RedisCommandKind::PubsubNumsub
+      | RedisCommandKind::PubsubShardchannels
+      | RedisCommandKind::PubsubShardnumsub => "PUBSUB",
       RedisCommandKind::_AuthAllCluster => "AUTH",
       RedisCommandKind::_HelloAllCluster(_) => "HELLO",
+      RedisCommandKind::_ClientTrackingCluster => "CLIENT",
       RedisCommandKind::_Custom(ref kind) => return kind.cmd.clone(),
     };
 
@@ -1181,7 +1209,6 @@ impl RedisCommandKind {
       RedisCommandKind::ClientInfo => "INFO",
       RedisCommandKind::ClientKill => "KILL",
       RedisCommandKind::ClientList => "LIST",
-      RedisCommandKind::ClientGetRedir => "GETREDIR",
       RedisCommandKind::ClientGetName => "GETNAME",
       RedisCommandKind::ClientPause => "PAUSE",
       RedisCommandKind::ClientUnpause => "UNPAUSE",
@@ -1190,6 +1217,10 @@ impl RedisCommandKind {
       RedisCommandKind::ClientSetname => "SETNAME",
       RedisCommandKind::ConfigGet => "GET",
       RedisCommandKind::ConfigRewrite => "REWRITE",
+      RedisCommandKind::ClientGetRedir => "GETREDIR",
+      RedisCommandKind::ClientTracking => "TRACKING",
+      RedisCommandKind::ClientTrackingInfo => "TRACKINGINFO",
+      RedisCommandKind::ClientCaching => "CACHING",
       RedisCommandKind::ConfigSet => "SET",
       RedisCommandKind::ConfigResetStat => "RESETSTAT",
       RedisCommandKind::MemoryDoctor => "DOCTOR",
@@ -1214,10 +1245,16 @@ impl RedisCommandKind {
       RedisCommandKind::FunctionLoad => "LOAD",
       RedisCommandKind::FunctionRestore => "RESTORE",
       RedisCommandKind::FunctionStats => "STATS",
+      RedisCommandKind::PubsubChannels => "CHANNELS",
+      RedisCommandKind::PubsubNumpat => "NUMPAT",
+      RedisCommandKind::PubsubNumsub => "NUMSUB",
+      RedisCommandKind::PubsubShardchannels => "SHARDCHANNELS",
+      RedisCommandKind::PubsubShardnumsub => "SHARDNUMSUB",
       RedisCommandKind::_FunctionLoadCluster => "LOAD",
       RedisCommandKind::_FunctionFlushCluster => "FLUSH",
       RedisCommandKind::_FunctionDeleteCluster => "DELETE",
       RedisCommandKind::_FunctionRestoreCluster => "RESTORE",
+      RedisCommandKind::_ClientTrackingCluster => "TRACKING",
       _ => return None,
     };
 
@@ -1227,10 +1264,6 @@ impl RedisCommandKind {
   pub fn use_random_cluster_node(&self) -> bool {
     match self {
       RedisCommandKind::Publish
-      | RedisCommandKind::Subscribe
-      | RedisCommandKind::Unsubscribe
-      | RedisCommandKind::Psubscribe
-      | RedisCommandKind::Punsubscribe
       | RedisCommandKind::Ping
       | RedisCommandKind::Info
       | RedisCommandKind::Scan
@@ -1253,7 +1286,8 @@ impl RedisCommandKind {
       | RedisCommandKind::Fcall
       | RedisCommandKind::FcallRO
       | RedisCommandKind::Wait => true,
-      // can be changed by the BLOCKING args
+      // default is false, but can be changed by the BLOCKING args. the RedisCommand::can_pipeline function checks the
+      // args too.
       RedisCommandKind::Xread | RedisCommandKind::Xreadgroup => false,
       RedisCommandKind::_Custom(ref kind) => kind.is_blocking,
       _ => false,
@@ -1267,6 +1301,7 @@ impl RedisCommandKind {
       | RedisCommandKind::_ScriptFlushCluster
       | RedisCommandKind::_ScriptKillCluster
       | RedisCommandKind::_HelloAllCluster(_)
+      | RedisCommandKind::_ClientTrackingCluster
       | RedisCommandKind::_ScriptLoadCluster
       | RedisCommandKind::_FunctionFlushCluster
       | RedisCommandKind::_FunctionDeleteCluster
@@ -1300,7 +1335,9 @@ impl RedisCommandKind {
     } else {
       match self {
         // make it easier to handle multiple potentially out-of-band responses
-        RedisCommandKind::Psubscribe
+        RedisCommandKind::Subscribe
+        | RedisCommandKind::Unsubscribe
+        | RedisCommandKind::Psubscribe
         | RedisCommandKind::Punsubscribe
         | RedisCommandKind::Ssubscribe
         | RedisCommandKind::Sunsubscribe
@@ -1356,6 +1393,8 @@ pub struct RedisCommand {
   pub network_start:     Option<Instant>,
   /// Whether to route the command to a replica, if possible.
   pub use_replica:       bool,
+  /// Only send the command to the provided server.
+  pub cluster_node:      Option<Server>,
   /// A timestamp of when the command was first created from the public interface.
   #[cfg(feature = "metrics")]
   pub created:           Instant,
@@ -1365,6 +1404,9 @@ pub struct RedisCommand {
   /// A counter to differentiate unique commands.
   #[cfg(feature = "debug-ids")]
   pub counter:           usize,
+  /// Whether to send a `CLIENT CACHING yes|no` before the command.
+  #[cfg(feature = "client-tracking")]
+  pub caching:           Option<bool>,
 }
 
 impl Drop for RedisCommand {
@@ -1416,13 +1458,16 @@ impl From<(RedisCommandKind, Vec<RedisValue>)> for RedisCommand {
       skip_backpressure: false,
       transaction_id: None,
       use_replica: false,
+      cluster_node: None,
+      network_start: None,
       #[cfg(feature = "metrics")]
       created: Instant::now(),
-      network_start: None,
       #[cfg(feature = "partial-tracing")]
       traces: CommandTraces::default(),
       #[cfg(feature = "debug-ids")]
       counter: command_counter(),
+      #[cfg(feature = "client-tracking")]
+      caching: None,
     }
   }
 }
@@ -1441,13 +1486,16 @@ impl From<(RedisCommandKind, Vec<RedisValue>, ResponseSender)> for RedisCommand 
       skip_backpressure: false,
       transaction_id: None,
       use_replica: false,
+      cluster_node: None,
+      network_start: None,
       #[cfg(feature = "metrics")]
       created: Instant::now(),
-      network_start: None,
       #[cfg(feature = "partial-tracing")]
       traces: CommandTraces::default(),
       #[cfg(feature = "debug-ids")]
       counter: command_counter(),
+      #[cfg(feature = "client-tracking")]
+      caching: None,
     }
   }
 }
@@ -1466,13 +1514,16 @@ impl From<(RedisCommandKind, Vec<RedisValue>, ResponseKind)> for RedisCommand {
       skip_backpressure: false,
       transaction_id: None,
       use_replica: false,
+      network_start: None,
+      cluster_node: None,
       #[cfg(feature = "metrics")]
       created: Instant::now(),
-      network_start: None,
       #[cfg(feature = "partial-tracing")]
       traces: CommandTraces::default(),
       #[cfg(feature = "debug-ids")]
       counter: command_counter(),
+      #[cfg(feature = "client-tracking")]
+      caching: None,
     }
   }
 }
@@ -1492,37 +1543,43 @@ impl RedisCommand {
       skip_backpressure: false,
       transaction_id: None,
       use_replica: false,
+      cluster_node: None,
+      network_start: None,
       #[cfg(feature = "metrics")]
       created: Instant::now(),
-      network_start: None,
       #[cfg(feature = "partial-tracing")]
       traces: CommandTraces::default(),
       #[cfg(feature = "debug-ids")]
       counter: command_counter(),
+      #[cfg(feature = "client-tracking")]
+      caching: None,
     }
   }
 
   /// Create a new empty `ASKING` command.
   pub fn new_asking(hash_slot: u16) -> Self {
     RedisCommand {
-      kind:                                       RedisCommandKind::Asking,
-      arguments:                                  Vec::new(),
-      response:                                   ResponseKind::Skip,
-      hasher:                                     ClusterHash::Custom(hash_slot),
-      timed_out:                                  Arc::new(AtomicBool::new(false)),
-      router_tx:                                  Arc::new(Mutex::new(None)),
-      attempted:                                  0,
-      can_pipeline:                               false,
-      skip_backpressure:                          true,
-      transaction_id:                             None,
-      use_replica:                                false,
+      kind:                                        RedisCommandKind::Asking,
+      arguments:                                   Vec::new(),
+      response:                                    ResponseKind::Skip,
+      hasher:                                      ClusterHash::Custom(hash_slot),
+      timed_out:                                   Arc::new(AtomicBool::new(false)),
+      router_tx:                                   Arc::new(Mutex::new(None)),
+      attempted:                                   0,
+      can_pipeline:                                false,
+      skip_backpressure:                           true,
+      transaction_id:                              None,
+      use_replica:                                 false,
+      cluster_node:                                None,
+      network_start:                               None,
       #[cfg(feature = "metrics")]
-      created:                                    Instant::now(),
-      network_start:                              None,
+      created:                                     Instant::now(),
       #[cfg(feature = "partial-tracing")]
-      traces:                                     CommandTraces::default(),
+      traces:                                      CommandTraces::default(),
       #[cfg(feature = "debug-ids")]
-      counter:                                    command_counter(),
+      counter:                                     command_counter(),
+      #[cfg(feature = "client-tracking")]
+      caching:                                     None,
     }
   }
 
@@ -1628,7 +1685,7 @@ impl RedisCommand {
   pub fn respond_to_router(&self, inner: &Arc<RedisClientInner>, cmd: RouterResponse) {
     if let Some(tx) = self.router_tx.lock().take() {
       if tx.send(cmd).is_err() {
-        _warn!(inner, "Failed to unblock router loop.");
+        _debug!(inner, "Failed to unblock router loop.");
       }
     }
   }
@@ -1657,6 +1714,7 @@ impl RedisCommand {
       can_pipeline: self.can_pipeline,
       skip_backpressure: self.skip_backpressure,
       router_tx: self.router_tx.clone(),
+      cluster_node: self.cluster_node.clone(),
       response,
       use_replica: self.use_replica,
       #[cfg(feature = "metrics")]
@@ -1666,6 +1724,8 @@ impl RedisCommand {
       traces: CommandTraces::default(),
       #[cfg(feature = "debug-ids")]
       counter: self.counter,
+      #[cfg(feature = "client-tracking")]
+      caching: self.caching.clone(),
     }
   }